OXDEV-7733 Move GreetingMessageService to Greeting domain
