+++ src/Greeting/Service/GreetingMessageService.php
@@ -7,12 +7,13 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\ModuleTemplate\Service;
+namespace OxidEsales\ModuleTemplate\Greeting\Service;
 
 use OxidEsales\Eshop\Application\Model\User as EshopModelUser;
 use OxidEsales\Eshop\Core\Request as EshopRequest;
 use OxidEsales\ModuleTemplate\Core\Module as ModuleCore;
 use OxidEsales\ModuleTemplate\Model\User as TemplateModelUser;
+use OxidEsales\ModuleTemplate\Service\ModuleSettingsServiceInterface;
 
 class GreetingMessageService implements GreetingMessageServiceInterface
 {
