+++ src/Greeting/Service/GreetingMessageServiceInterface.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\ModuleTemplate\Service;
+namespace OxidEsales\ModuleTemplate\Greeting\Service;
 
 use OxidEsales\Eshop\Application\Model\User as EshopModelUser;
