+++ src/Service/TrackerService.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace OxidEsales\ModuleTemplate\Service;
 
 use OxidEsales\Eshop\Application\Model\User as EshopModelUser;
-use OxidEsales\ModuleTemplate\Greeting\Repository as GreetingRepository;
+use OxidEsales\ModuleTemplate\Greeting\Repository\GreetingRepositoryInterface;
 use OxidEsales\ModuleTemplate\Model\User as ModelUser;
 use OxidEsales\ModuleTemplate\Tracker\Repository as TrackerRepository;
 
@@ -21,7 +21,7 @@ class TrackerService implements TrackerServiceInterface
 {
     public function __construct(
         private TrackerRepository $repository,
-        private GreetingRepository $greetingRepository,
+        private GreetingRepositoryInterface $greetingRepository,
     ) {
     }
